Fix Share->share
Since "Share on twitter" is not  a new sentence (it comes after a
comma (,), not a period (.)), it should start with a lower-case
letter, not a capital one.

This patch was accepted on the ap2016 branch, but unfortunately it
was merged only after the ap2017 branch was branched out of it.